Replace code with ChibiOS macro rewrite and correct return type
- ChibiOS macro STMS(n) will overflow when doing the math with uint32_t

+++ src/CLR/Include/WireProtocol.h
@@ -116,7 +116,7 @@ struct WP_Message
         static const int CompletePayload  = 6;
     };
 
-    static const UINT32 c_PayloadTimeout = 6 * 1000; // 6 secs (from milliseconds time)
+    static const UINT64 c_PayloadTimeout = 6 * 1000; // 6 secs (from milliseconds time)
 
     WP_Controller* m_parent;
     WP_Packet      m_header;
